Speed up tooltip display &  refactor collapsable
