+++ src/const/sizes.js
@@ -8,6 +8,13 @@ export const DEVICES = {
   xl: 1281,
 };
 
+// MUI breakpoints
+// xs: 0px
+// sm: 600px
+// md: 900px
+// lg: 1200px
+// xl: 1536px
+
 export const MAP_HEIGHT = {
   default: '30vh',
   upSmall: '35vh',
