+++ src/pages/Faq.js
@@ -1,41 +1,13 @@
-import { useRef, useCallback, useEffect, useState } from 'react';
+import { useRef, useEffect, useState } from 'react';
 import { t } from 'i18next';
-import Tooltip from '@mui/material/Tooltip';
 import { Loader } from 'components/Shared';
+import { Collapsable } from 'components/Shared/Collapsable';
 import * as Styled from './styles/Markdown';
 
 const Faq = function Faq() {
   const lng = localStorage.getItem('i18nextLng') || 'sl';
   const faqRef = useRef();
   const [response, setResponse] = useState();
-
-  // scroll element to when link with hash is passed
-  const scroll = useCallback(node => {
-    if (node !== null && node.id === window.location.hash.substr(1)) {
-      // eslint-disable-next-line no-param-reassign
-      node.open = true;
-      window.scrollTo({
-        top: node.getBoundingClientRect().top,
-        behavior: 'smooth',
-      });
-    }
-  }, []);
-
-  // copy url of the definition
-  const handleCopy = e => {
-    const element = e.currentTarget;
-    let text = `${window.location.href}#${element.nextSibling.id}`;
-    if (window.location.hash !== '') {
-      text = `${window.location.href.split('#')[0]}#${element.nextSibling.id}`;
-    }
-    navigator.clipboard.writeText(text);
-    element.className = 'icon check';
-    element.title = t('copied');
-    setTimeout(() => {
-      element.className = 'icon copy';
-      element.title = t('copy');
-    }, 2000);
-  };
 
   useEffect(() => {
     document.body.style.overflow = 'auto';
@@ -50,19 +22,20 @@ const Faq = function Faq() {
       });
   }, [lng]);
 
-  // append attribute 'title' with glossary definition to each element with attribute 'data-term'
   useEffect(() => {
     if (faqRef.current) {
+      // append 'data-definition' attribute with glossary definition to each element with attribute 'data-term'
       faqRef.current.querySelectorAll('span[data-term]').forEach(el => {
         // eslint-disable-next-line no-restricted-syntax
         for (const term of response.glossary) {
           if (term.slug === el.getAttribute('data-term')) {
-            el.setAttribute('title', term.definition.replace(/<[^>]*>?/gm, ''));
+            // TODO: is it possible to render term.definition as Markdown text?
+            el.setAttribute('data-definition', term.definition.replace(/<[^>]*>?/gm, ''));
             el.setAttribute('tabindex', 0);
           }
         }
       });
-      // append attribute target="blank" to all external links
+      // append attribute target="_blank" to all external links
       if (faqRef.current) {
         faqRef.current.querySelectorAll('a').forEach(el => {
           if (/^(https?:)?\/\//.test(el.getAttribute('href'))) {
@@ -73,43 +46,37 @@ const Faq = function Faq() {
     }
   }, [faqRef, response]);
 
-  if (response == null) {
+  if (!response) {
     return <Loader.Center />;
   }
 
   return (
-    <Styled.CustomContainer className="custom-container">
-      <Styled.StaticPageWrapper className="static-page-wrapper" ref={faqRef}>
+    <Styled.CustomContainer>
+      <Styled.StaticPageWrapper ref={faqRef}>
         <h1>{t('faq.title')}</h1>
         <p>{t('faq.description')}</p>
         <br />
         {response.faq.map((faq, index) => {
           const collapsableKey = `collapsable-faq-${index}`;
           return (
-            <Styled.Collapsable className="collapsable" key={collapsableKey}>
-              <Tooltip title={<div>{t('copy')}</div>} placement="top">
-                <Styled.IconWrapper className="icon copy" alt={t('copy')} onClick={handleCopy} />
-              </Tooltip>
-              <Styled.Details id={faq.slug} ref={scroll}>
-                <Styled.Summary>{faq.question}</Styled.Summary>
-                <Styled.Markdown key={`${faq.slug}-md`}>{faq.answer}</Styled.Markdown>
-              </Styled.Details>
-            </Styled.Collapsable>
+            <Collapsable
+              key={collapsableKey}
+              id={faq.slug}
+              question={faq.question}
+              answer={faq.answer}
+            />
           );
         })}
         <h2>{t('faq.glossary')}</h2>
         {response.glossary.map((glossary, index) => {
           const collapsableKey = `collapsable-glossary-${index}`;
           return (
-            <Styled.Collapsable className="collapsable" key={collapsableKey}>
-              <Tooltip title={<div>{t('copy')}</div>} placement="top">
-                <Styled.IconWrapper className="icon copy" alt={t('copy')} onClick={handleCopy} />
-              </Tooltip>
-              <Styled.Details id={glossary.slug} ref={scroll}>
-                <Styled.Summary>{glossary.term}</Styled.Summary>
-                <Styled.Markdown key={`${glossary.slug}-md`}>{glossary.definition}</Styled.Markdown>
-              </Styled.Details>
-            </Styled.Collapsable>
+            <Collapsable
+              key={collapsableKey}
+              id={glossary.slug}
+              question={glossary.term}
+              answer={glossary.definition}
+            />
           );
         })}
       </Styled.StaticPageWrapper>
