--- src/pages/style.scss
@@ -1,437 +1,0 @@
-$element-box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.15);
-$text-c: rgba(0, 0, 0, 0.7);
-$green: #95C83F;
-
-.custom-container {
-  margin: 24px auto 0 auto;
-  max-width: 730px;
-
-  @media only screen and (min-width: 768px) {
-    margin: 48px auto 65px auto;
-    box-shadow: $element-box-shadow;
-  }
-}
-
-.static-page-wrapper {
-  padding: 32px 17px 27px 17px;
-  background: #fff;
-  font-size: 14px;
-  color: rgba(0,0,0,.7);
-  line-height: 1.7;
-
-  @media only screen and (min-width: 768px) {
-    padding: 32px 32px 27px 32px;
-  }
-
-  h1 {
-    margin-bottom: 32px;
-  }
-
-  .dropdown + h2,
-  .dropdown + h3,
-  h1 + h2 {
-    margin-top: 64px;
-  }
-
-  h2,
-  h3,
-  h4 {
-    margin-bottom: 24px;
-  }
-
-  h1 {
-    font-size: 28px;
-  }
-
-  h2 {
-    font-size: 21px;
-  }
-
-  h3 {
-    font-size: 18px;
-  }
-
-  p:not(:last-of-type) {
-    margin-bottom: 28px;
-  }
-
-  ul, ol {
-    padding-inline-start: 40px;
-    margin-bottom: 1rem;
-  }
-
-  //subtitle
-  h1 + p > em {
-    display: block;
-    font-size: 16px;
-    font-style: italic;
-    color: rgba(0, 0, 0, 0.8);
-    font-weight: 400;
-    line-height: 1.7;
-    margin-bottom: 48px;
-
-    a {
-      font-size: 16px;
-    }
-  }
-
-  p,
-  a,
-  span,
-  strong {
-    font-size: 14px;
-    color: $text-c;
-    line-height: 1.7;
-  }
-
-  * + h1,
-  * + h2,
-  * + h3,
-  * + table {
-    margin-top: 48px;
-  }
-
-  tr + tr {
-    margin-top: 27px;
-  }
-
-  table {
-    width: 100%;
-    table-layout: fixed;
-    text-align: left;
-    margin-bottom: 1rem;
-  }
-
-  table {
-    table-layout: fixed;
-    text-align: left;
-    td {
-      padding: 15px 0;
-      width: 50%;
-      border-top: 1px solid rgba(0, 0, 0, 0.45);
-    }
-  }
-
-  //dropdown HTML in MD
-  h1 + details,
-  h2 + details,
-  h3 + details {
-    margin-top: 48px;
-  }
-
-  details,
-  .dropdown {
-    margin-bottom: 24px;
-
-    & + details,
-    & + .dropdown {
-      border-top: 1px solid #dedede;
-      padding-top: 24px;
-    }
-
-    summary,
-    .dd-title {
-      cursor: pointer;
-      user-select: none;
-      font-weight: bold;
-      font-stretch: normal;
-      font-style: normal;
-      line-height: 1.71;
-      color: rgba(0, 0, 0, 0.75);
-      position: relative;
-      padding-right: 18%;
-
-      @media only screen and (min-width: 768px) {
-        padding-right: 10%;
-      }
-      
-      &:after {
-        content: url('../assets/expand-dd.svg');
-        display: block;
-        position: absolute;
-        right: 0;
-        top: 0;
-      }
-
-      &:focus {
-        outline: none;
-      }
-    }
-  }
-
-  .dropdown .dd-content {
-    display: none;
-    margin: 2px 0;
-    padding-top: 12px;
-    width: 90%;
-    position: relative;
-  }
-
-  details > *:not(summary) {
-    position: relative;
-    display: none;
-    width: 90%;
-  }
-
-  details > *:nth-child(2) {
-    margin-top: 2px;
-    padding-top: 12px;
-  }
-
-  details[open] {
-    & > *:not(summary) {
-      display: block;
-      animation: show-dd 0.5s ease-out;
-    }
-
-    summary {
-      &:after {
-        content: url('../assets/close-dd.svg');
-      }
-    }
-  }
-
-  .dropdown.dd-show {
-    .dd-content {
-      display: block;
-      animation: show-dd 0.5s ease-out;
-    }
-
-    .dd-title {
-      &:after {
-        content: url('../assets/close-dd.svg');
-      }
-    }
-  }
-
-  @keyframes show-dd {
-    from {
-      transform: translateY(-8px);
-      opacity: 0.1;
-    }
-    to {
-      transform: translateY(0px);
-      opacity: 1;
-    }
-  }
-
-  .img-link {
-    display: block;
-    box-shadow: none;
-    margin-bottom: 24px;
-
-    &:hover {
-      box-shadow: none;
-    }
-
-    img {
-      width: 100%;
-        &.rightThumbnail {
-          @media only screen and (min-width: 480px) {
-            width: 50%;
-            float:right;
-            margin: 0 0 20px 20px;
-          }
-        }
-
-    }
-  }
-
-  div.logos {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    flex-grow: 1;
-    margin-bottom: 12px;
-
-    a {
-      display: inline-block;
-      flex: 1 1 0;
-      margin-bottom: 12px;
-      max-width: 230px;
-      min-width: 200px;
-      box-shadow: none;
-      text-decoration: none;
-
-      &:hover {
-        box-shadow: none;
-        text-decoration: none;
-      }
-
-      img {
-        padding: 10px;
-        width: 100%;
-      }
-    }
-  }
-
-  iframe.youtube {
-    width:320px;
-    height:180px;
-    max-width:100%;
-  }
-  
-  iframe.observable {
-    width:660px;
-    height:67vw;
-    max-width:100%;
-    max-height:440px;
-    min-width:400px;
-    min-height:300px
-  }
-}
-
-details summary::-webkit-details-marker {
-  display: none;
-}
-
-.static-page-wrapper,
-.footnote,
-.link {
-  a {
-    font-weight: 600;
-    transition: all 0.35s ease-in-out;
-    box-shadow: inset 0 -1px 0 white, inset 0 -4px $green;
-    text-decoration: none;
-    color: rgba(0, 0, 0, 0.8);
-
-    &:hover {
-      text-decoration: none;
-      color: rgba(0, 0, 0, 0.8);
-      font-weight: 600;
-      box-shadow: inset 0 -1px 0 white, inset 0 -20px $green;
-    }
-
-    strong {
-      font-weight: 600;
-    }
-  }
-}
-
-/////////////////////////////
-//          FAQ            //   TODO: could someone please convert it to MUI?
-/////////////////////////////
-
-details {
-  margin-bottom: 24px;
-
-  & + details {
-    border-top: 1px solid #dedede;
-    padding-top: 24px;
-  }
-
-  summary {
-    cursor: pointer;
-    user-select: none;
-    font-weight: bold;
-    font-stretch: normal;
-    font-style: normal;
-    line-height: 1.71;
-    color: rgba(0, 0, 0, 0.75);
-    position: relative;
-    padding-right: 18%;
-
-    @media only screen and (min-width: 768px) {
-      padding-right: 10%;
-    }
-
-    summary::-webkit-details-marker {
-        display: none;
-    }
-
-    &:after {
-      content: url("../assets/expand-dd.svg");
-      display: block;
-      position: absolute;
-      right: 0;
-      top: 0;
-    }
-
-    &:focus {
-      outline: none;
-    }
-  }
-}
-
-details > *:not(summary) {
-  position: relative;
-  display: none;
-  width: 90%;
-}
-
-details > *:nth-child(2) {
-  margin-top: 2px;
-  padding-top: 12px;
-}
-
-details[open] {
-  & > *:not(summary) {
-    display: block;
-    animation: show-dd 0.5s ease-out;
-  }
-
-  summary {
-    &:after {
-      content: url("../assets/close-dd.svg");
-    }
-  }
-}
-
-.collapsable {
-  position: relative;
-}
-.icon {
-  width: 28px;
-  cursor: pointer;
-  padding: 5px;
-  position: absolute;
-  top: -3px;
-  right: 30px;
-  z-index: 10;
-  &.copy {
-    content: url("../assets/icon-copy.svg");
-  }
-  &.check {
-    content: url("../assets/check.svg");
-  }
-}
-
-span[data-term] {
-  cursor: help;
-  font-weight: 600;
-  transition: all 0.35s ease-in-out;
-  box-shadow: inset 0 -1px 0 white, inset 0 -4px $green;
-  text-decoration: none;
-  color: rgba(0, 0, 0, 0.8);
-
-  &:hover {
-    text-decoration: none;
-    color: rgba(0, 0, 0, 0.8);
-    font-weight: 600;
-    box-shadow: inset 0 -1px 0 white, inset 0 -20px $green;
-  }
-}
-@media (pointer: coarse), (hover: none) {
-  span[data-term] {
-    position: relative;
-    display: inline-flex;
-    justify-content: center;
-    outline: none;
-  }
-  span[data-term]:focus::after {
-    content: attr(title);
-    position: absolute;
-    top: 90%;
-    width: 200px;
-    background-color: #414040;
-    color: white;
-    border: 1px solid;
-    padding: 3px 6px;
-    margin: 10px;
-    font-size: 10px;
-    font-weight: 200;
-    line-height: 1.4;
-    z-index: 1;
-  }
-}
