+++ src/pages/styles/Markdown.js
@@ -1,26 +1,10 @@
 import { styled } from '@mui/material/styles';
-import MarkdownBase from 'markdown-to-jsx';
-import iconExpand from 'assets/expand-dd.svg';
-import iconClose from 'assets/close-dd.svg';
-import iconCheck from 'assets/check.svg';
-import iconCopy from 'assets/icon-copy.svg';
 
-export const IconWrapper = styled('span')(() => ({
-  width: '28px',
-  cursor: 'pointer',
-  padding: '5px',
-  position: 'absolute',
-  top: '-3px',
-  right: '30px',
-  zIndex: 10,
-}));
-
-// Names of CustomContainer and StaticPageWrapper matches classes from 'style.scss'
 export const CustomContainer = styled('main')(({ theme }) => ({
   margin: '24px auto 0 auto',
   maxWidth: '730px',
   background: theme.palette.common.white,
-  '@media only screen and (min-width: 768px)': {
+  [theme.breakpoints.up('md')]: {
     margin: '48px auto 65px auto',
     boxShadow: theme.MD.elementBoxShadow,
   },
@@ -30,7 +14,7 @@ export const StaticPageWrapper = styled('div')(({ theme }) => ({
   fontSize: '14px',
   padding: '32px 17px 27px 17px',
 
-  '@media only screen and (min-width: 768px)': {
+  [theme.breakpoints.up('md')]: {
     margin: '32px 32px 27px 32px',
   },
 
@@ -87,30 +71,27 @@ export const StaticPageWrapper = styled('div')(({ theme }) => ({
   },
 
   'span[data-term]': {
+    position: 'relative',
+    display: 'inline-flex',
+    justifyContent: 'center',
+    outline: 'none',
     cursor: 'help',
     fontWeight: 600,
     transition: 'all 0.35s ease-in-out',
-    boxShadow: theme.MD.dataTermBoxShadow,
-    textDecoration: 'none',
     color: 'rgba(0, 0, 0, 0.8)',
 
-    '&:hover ': {
-      textDecoration: 'none',
-      color: 'rgba(0, 0, 0, 0.8)',
-      fontWeight: 600,
-      boxShadow: theme.MD.dataTermBoxShadowHover,
+    '&:before ': {
+      content: '""',
+      width: '100%',
+      height: '100%',
+      position: 'absolute',
+      top: '-2px',
+      borderBottom: `2px ${theme.customColors.brand} dotted`,
+      zIndex: 1,
     },
-  },
 
-  '@media (pointer: coarse), (hover: none)': {
-    'span[data-term]': {
-      position: 'relative',
-      display: 'inline-flex',
-      justifyContent: 'center',
-      outline: 'none',
-    },
-    'span[data-term]:focus::after': {
-      content: 'attr(title)',
+    '&:hover::after': {
+      content: 'attr(data-definition)',
       position: 'absolute',
       top: '90%',
       width: '100px',
@@ -122,7 +103,7 @@ export const StaticPageWrapper = styled('div')(({ theme }) => ({
       fontSize: '10px',
       fontWeight: 200,
       lineHeight: 1.4,
-      zIndex: 1,
+      zIndex: 2,
     },
   },
 
@@ -139,92 +120,5 @@ export const StaticPageWrapper = styled('div')(({ theme }) => ({
       width: '50%',
       borderTop: `1px solid ${theme.MD.tableTdBorder}`,
     },
-  },
-
-  '@keyframes show-dd': {
-    from: {
-      transform: 'translateY(-8px)',
-      opacity: 0.1,
-    },
-    to: {
-      transform: 'translateY(0px)',
-      opacity: 1,
-    },
-  },
-}));
-
-export const Collapsable = styled('div')(() => ({
-  position: 'relative',
-  fontSize: 'inherit',
-  '.icon': {
-    width: '28px',
-    cursor: 'pointer',
-    padding: '5px',
-    position: 'absolute',
-    top: '-3px',
-    right: '30px',
-    zIndex: 10,
-    '&.copy': {
-      content: `url(${iconCopy})`,
-    },
-    '&.check': {
-      content: `url(${iconCheck})`,
-    },
-  },
-}));
-
-// !todo remove Markdown after everything is merged and use in MarkdownBase in  FAQ and About page
-export const Markdown = styled(MarkdownBase)(() => ({}));
-
-export const Details = styled('details')(() => ({
-  fontSize: 'inherit',
-  marginBottom: '24px',
-  '> *:not(summary)': {
-    position: 'relative',
-    display: 'none',
-    width: '90%',
-  },
-  '> *:nth-child(2)': {
-    marginTop: '2px',
-    paddingTop: '12px',
-  },
-  '&[open]': {
-    '> *:not(summary)': {
-      display: 'block',
-      animation: 'show-dd 0.5s ease-out',
-    },
-    'summary::after': {
-      content: `url(${iconClose})`,
-    },
-  },
-}));
-
-export const Summary = styled('summary')(({ theme }) => ({
-  cursor: 'pointer',
-  userSelect: 'none',
-  fontSize: 'inherit',
-  fontWeight: 'bold',
-  fontStretch: 'normal',
-  fontStyle: 'normal',
-  lineHeight: 1.71,
-  color: theme.MD.summaryColor,
-  position: 'relative',
-  paddingRight: '18%',
-  '@media only screen and (min-width: 768px)': {
-    paddingRight: '10%',
-  },
-  '::marker': {
-    display: 'none',
-    content: '""',
-  },
-  '::after': {
-    content: `url(${iconExpand})`,
-    display: 'block',
-    position: 'absolute',
-    right: 0,
-    top: 0,
-  },
-  '$:focus': {
-    outline: 'none',
   },
 }));
